[bazel] Bump `rules_closure` and google closure libary to latest release (#16755)
* [bazel] Bump `rules_closure` and google closure libary to latest release

The Google Closure Library is the last public release, and it seems
unlikely we'll ever update it again.

The new version of `rules_closure` brings a new version of the closure
compiler. This is stricter than the previous compiler, and necessitated
an update of the Google Closure Library. However, this new version of
the library had removed some key functions we relied upon. We've had
to rework both our own code and the vendored version of closure to get
it to work with the latest closure compiler.

* cp: set a flag to allow protobuf to compile on windows

* cp: use a wrapper to avoid line length issues on windows