[java] Fix asserts for maps and sets (#16808)
* Fix asserts for Sets and Maps

The problem is that AssertJ's method `isEqualTo` doesn't always work for Sets and Maps because it assumes strict order of elements:
1. `assertThat(MAP).isEqualTo(Map.of(...))`
2. `assertThat(SET).isEqualTo(Set.of(...))`

In both cases, we have to replace `isEqualTo` by `containsExactlyInAnyOrderEntriesOf` or similar method.

* Simplify creating lists in ByChainedTest

After the simplification, it became obvious that some of ByChainedTest tests don't really test anything. They create a bunch of lists (e.g. `elems345`) which are not used. And only the empty list of elements is loaded and asserted.

Seems that this complex/misleading test logic in ByChainedTest was
introduced in commit c4861b86 (some "patch from BenChambers"). Needs to be reviewed and fixed.

* slightly improve PDF check in BrowsingContextTest

now we check that the resulting content at least _looks like PDF_.